api.js: - reworked;
~/services/basecrud.js:
- added;
